+++ Control/Monad/Identity.hs
@@ -35,13 +35,10 @@ version of that monad.
 module Control.Monad.Identity (
     module Data.Functor.Identity,
     module Control.Monad.Trans.Identity,
-
-    module Control.Monad,
-    module Control.Monad.Fix,
    ) where
 
-import Data.Functor.Identity
+-- We do this so, because transformers-0.4.0.0
+-- https://hackage.haskell.org/package/transformers-0.4.0.0/docs/Data-Functor-Identity.html
+-- doesn't have runIdentity as a field name.
+import Data.Functor.Identity (Identity(Identity), runIdentity)
 import Control.Monad.Trans.Identity
-
-import Control.Monad
-import Control.Monad.Fix